+++ CHANGELOG.md
@@ -21,6 +21,10 @@ SPDX-License-Identifier: GPL-3.0-or-later
 - Add option `-o` (or `--output-format`) to customize output format, options `-e` and `-m` are now aliases on this option
 - Add official pre-commit hook
 
+### Fixed
+
+- Do not display empty line when no misspelled word is found with option `-m` / `--only-misspelled` / `--output-format misspelled`
+
 ## Version 4.1.0 (2024-10-23)
 
 ### Changed